+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Added plot comparing countries from the day where the total number of confirmed cases was above 1000 (#39).
 - Added growth factor and Death rate plots (#5)
+- Added new cases to map (#56)
 
 ### Covid19 1.1.3 (2020-03-30)